+++ trunk/castor/src/main/org/exolab/castor/types/Month.java
@@ -239,7 +239,7 @@ public class Month extends TimePeriod {
         // Set the time zone
         if ( !isUTC() ) {
             int offset = 0;
-            offset = (int) ( (this.getZoneMinute() + this.getZoneHour()*60)*60*1000);
+            offset = ( (this.getZoneMinute() + this.getZoneHour()*60)*60*1000);
             offset = isZoneNegative() ? -offset : offset;
             timeZone.setRawOffset(offset);
             timeZone.setID(TimeZone.getAvailableIDs(offset)[0]);
