+++ trunk/castor/src/main/org/exolab/castor/types/RecurringDuration.java
@@ -360,7 +360,7 @@ public class RecurringDuration extends RecurringDurationBase{
         // Set the time zone
         if ( !isUTC() ) {
             int offset = 0;
-            offset = (int) ( (this.getZoneMinute() + this.getZoneHour()*60)*60*1000);
+            offset = ( (this.getZoneMinute() + this.getZoneHour()*60)*60*1000);
             offset = isZoneNegative() ? -offset : offset;
             timeZone.setRawOffset(offset);
             timeZone.setID(TimeZone.getAvailableIDs(offset)[0]);
